.deploy/knowledge_base/pyproject.toml
[project]
name = "rag-youtuber-kanilla"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "azure-functions>=1.24.0",
    "fastapi>=0.124.4",
    "google-generativeai>=0.8.5",
    "ipykernel>=7.1.0",
    "lancedb>=0.25.3",
    "langchain-community>=0.4.1",
    "openai>=1.50.0",
    "pandas>=2.3.3",
    "pydantic-ai>=1.32.0",
    "pydantic-settings>=2.12.0",
    "streamlit>=1.52.1",
    "uvicorn>=0.38.0",
]
